// frontend/src/app/core/model/model.d.ts
export interface ISubCategory {

    /** SubCategory id */
    id?: (number|null);

    /** SubCategory isDeleted */
    isDeleted?: (boolean|null);

    /** SubCategory isVisible */
    isVisible?: (boolean|null);

    /** SubCategory transactionAmount */
    transactionAmount?: (number|null);

    /** SubCategory name */
    name: string;

    /** SubCategory categoryId */
    categoryId: number;
}

/** Represents a SubCategory. */
export class SubCategory implements ISubCategory {

    /**
     * Constructs a new SubCategory.
     * @param [properties] Properties to set
     */
    constructor(properties?: ISubCategory);

    /** SubCategory id. */
    public id: number;

    /** SubCategory isDeleted. */
    public isDeleted: boolean;

    /** SubCategory isVisible. */
    public isVisible: boolean;

    /** SubCategory transactionAmount. */
    public transactionAmount: number;

    /** SubCategory name. */
    public name: string;

    /** SubCategory categoryId. */
    public categoryId: number;

    /**
     * Creates a new SubCategory instance using the specified properties.
     * @param [properties] Properties to set
     * @returns SubCategory instance
     */
    public static create(properties?: ISubCategory): SubCategory;

    /**
     * Encodes the specified SubCategory message. Does not implicitly {@link SubCategory.verify|verify} messages.
     * @param message SubCategory message or plain object to encode
     * @param [writer] Writer to encode to
     * @returns Writer
     */
    public static encode(message: ISubCategory, writer?: $protobuf.Writer): $protobuf.Writer;

    /**
     * Decodes a SubCategory message from the specified reader or buffer.
     * @param reader Reader or buffer to decode from
     * @param [length] Message length if known beforehand
     * @returns SubCategory
     * @throws {Error} If the payload is not a reader or valid buffer
     * @throws {$protobuf.util.ProtocolError} If required fields are missing
     */
    public static decode(reader: ($protobuf.Reader|Uint8Array), length?: number): SubCategory;

    /**
     * Creates a SubCategory message from a plain object. Also converts values to their respective internal types.
     * @param object Plain object
     * @returns SubCategory
     */
    public static fromObject(object: { [k: string]: any }): SubCategory;

    /**
     * Creates a plain object from a SubCategory message. Also converts values to other types if specified.
     * @param message SubCategory
     * @param [options] Conversion options
     * @returns Plain object
     */
    public static toObject(message: SubCategory, options?: $protobuf.IConversionOptions): { [k: string]: any };

    /**
     * Converts this SubCategory to JSON.
     * @returns JSON object
     */
    public toJSON(): { [k: string]: any };
}
// ...
export enum AppMode {
    FINANCE = 0,
    NOTES = 1
}